+++ _site/scripts/project_tier_table_generator.py
@@ -78,9 +78,9 @@ def create_repo_badges(username):
 
     repo_badge_template = f"""## Tier Badges
 
-Use this file as a template to gather and add badges to your project's `README.md` files.
-- Be sure to run the workflow to automatically update the badges with your username
-- Optionally, you can change the `href` attributes below to point to your project's repository by changing the username to your GitHub username.
+Use this file as a template to gather and add badges to your project's `README` files.
+- Be sure to run the workflow to automatically update the badges below with your username.
+- Prior to running the workflow, update the `project_tier_table_generator.py` script's `TIER_TABLE_URL` parameter to point to your project's **PRG** website/domain.
 
 ### Gold Project Badge